Switch all save* functions to (data, callback) signature + normalise to "_"'s
Rather than passing a load of parameters to the model save* functions,
now they are passed two parameters "data" and the callback.
To aid with this (and for other sanity reasons) all interfaces with the
model now use underscored variables (e.g. req.oauth.grantType -> grant_type)
Also included are a few fixes in tests etc.