Modified Authorization Grant to accept a GET with parameters in the Query String
The RFC Spec for OAuth 2 specifies that the Authorization Grant endpoint be
a HTTP GET.  POST is allowed to be supported, but GET must be implemented
according to the spec.
This change modifies authCodeGrant.js to look for the appropriate parameters
in either the query string or the body, so HTTP GET and HTTP POST are both
supported.

Added tests to test/authCodeGrant.js to test both POST and GET functionality.
Closes #56