Added ability for client to have multiple redirect URIs which are valid.
The OAuth 2 spec envisions that an individual client can have several redirect URIs that are valid for a given clientId.  This change allows the redirect_uri returned from the model to be either a string or an array.  Added test for an invalid URI when multiple URIs are returned.  Also added tests for valid URIs when either a string or an Array is returned from getClient.