Implements scope support
Resolves https://github.com/thomseddon/node-oauth2-server#49 by obtaining an
optional scope parameter from a grant request, which is passed to the model
for processing and storage. The authorization flow includes a call
to `checkScope` which receives the access token and the request object.

In order to support the access code grant flow, the scope param is
passed to `saveAuthCode` so that it may be persisted for a subsequent
token exchange.