Implements scope support
Resolves https://github.com/thomseddon/node-oauth2-server#49 by obtaining an
optional scope parameter from a grant request, which is passed to the model
for processing and storage. The authorization flow includes a call
to `checkScope` which receives the access token and the request object.

In order to support the access code grant flow, the scope param is
passed to `saveAuthCode` so that it may be persisted for a subsequent
token exchange.

Scoping is now implemented using middleware

Custom middleware allows for defining scope key requirements
on routers or single routes. If a scope requirement is set,
the model's checkScope() method is tasked with verifying whether
or not the access token possesses the required key(s).

The checkScope method now receives the entire bearerToken object

This change allows for simpler access to the scope information,
should the developer chose to include it in the getAccessToken
response.

Updated Readme and postgresql example

Added method descriptors for saveScope and checkScope to the Readme;
added a scope param to saveAuthCode, since the user-accepted string
must be passed along server-side to the token creation step.

Corresponding updates were made to the postgresql example, which
now includes a simple usage scenario with a scope string being stored
in the oauth_access_tokens table.

Documentation fixes and checkScope parameter order changed

Added optional scope argument to authorise middleware

In addition to the dedicated scope middleware, required scope may now
also be defined through the authorise middleware. Multiple scope
definitions cascade, meaning that it is possible to protect a router
instance via authorise, and add finer-grained requirements for
individual routes.

Fix parameter order in postgres model

Improve scope support in authorise middleware

Remove scope middleware

Improve scope support in grant middleware

Conflicts:
	lib/grant.js
	lib/oauth2server.js