Updated Readme and postgresql example
Added method descriptors for saveScope and checkScope to the Readme;
added a scope param to saveAuthCode, since the user-accepted string
must be passed along server-side to the token creation step.

Corresponding updates were made to the postgresql example, which
now includes a simple usage scenario with a scope string being stored
in the oauth_access_tokens table.