Scope improvements
The grant middleware now delegates scope string validation to the model,
allowing for proper error handling in case a client performs an invalid
request. Valid scope strings are thereafter passed to the model
in the `saveAccessToken` call; `saveScope` has been removed.

The `authoriseScope` method has been modified to accept a mixed `invalid`
argument, which should be falsey to indicate that the access token
possesses the required scope, and truthy (boolean or string) if not.
If a string is provided, it is used as error description for the
`invalid_scope` error. The error class has been extended with a default
"Invalid scope" message for boolean true cases.