Scope validation on granting
These changes renames the present `model.validateScope` to
`model.verifyScope` and introduces a new `model.validateScope` method
that is run when granting (except for the refresh token grant flow,
which inherits the previous token's scope string).

The new `validateScope` takes the user, client, and requested scope
string as input, and allows the model to perform arbitrarily complex
validations and/or sanitations (e.g. client A may not request the admin
scope, so filter it out) and potentially abort the grant flow with an
invalid scope error.